Correctly support associations in autotype data
Keepass2 allows for multiple window associations to be defined.
The XML file format just has multiple Association child nodes
in the AutoType data, making it a non obvious behaviour from looking
at an example file (which usually only has 1 window association defined)
This changes the behaviour.

While technically a breaking change, it is also a bug fix
for incorrectly handling the data.
The assumption is that few of the gokeepasslib users have used this
so far, which makes it a less intrusive breaking change.
Due to that I decided to just bump it from v3.2.1 to v3.2.2